+++ packages/engine/Specs/Scene/ScreenSpaceCameraControllerSpec.js
@@ -34,6 +34,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
   function MockScene(canvas, camera, ellipsoid) {
     this.canvas = canvas;
     this.camera = camera;
+    this.ellipsoid = ellipsoid;
     this.globe = undefined;
     this.verticalExaggeration = 1.0;
     this.verticalExaggerationRelativeHeight = 0.0;
@@ -271,7 +272,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   function setUp3DUnderground() {
     setUp3D();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
     scene.cameraUnderground = true;
     controller.enableCollisionDetection = false;
 
@@ -1198,7 +1199,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
   it("zooms in on an object in 3D", function () {
     setUp3D();
 
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();
 
@@ -1229,7 +1230,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
   it("zooms in on an object in 3D when transform is set", function () {
     setUp3D();
 
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();
 
@@ -1261,7 +1262,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("zoom in 3D to point 0,0", function () {
     setUp3D();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();
 
@@ -1482,10 +1483,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
     ).toEqualEpsilon(camera.up, CesiumMath.EPSILON14);
 
     const ray = new Ray(camera.positionWC, camera.directionWC);
-    const intersection = IntersectionTests.rayEllipsoid(
-      ray,
-      scene.mapProjection.ellipsoid
-    );
+    const intersection = IntersectionTests.rayEllipsoid(ray, scene.ellipsoid);
     expect(intersection).toBeDefined();
   });
 
@@ -1810,7 +1808,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("camera does not go below the terrain in 3D", function () {
     setUp3D();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();
 
@@ -1839,7 +1837,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("camera does not go below the terrain in CV", function () {
     setUpCV();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();
 
@@ -1868,7 +1866,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("camera does go below the terrain in 3D when collision detection is disabled", function () {
     setUp3D();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
     controller.enableCollisionDetection = false;
 
     updateController();
@@ -1897,7 +1895,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("camera does go below the terrain in CV when collision detection is disabled", function () {
     setUpCV();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
     controller.enableCollisionDetection = false;
 
     updateController();
@@ -1924,7 +1922,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("camera does not go below the terrain in 3D with the transform set", function () {
     setUp3D();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();
 
@@ -1953,7 +1951,7 @@ describe("Scene/ScreenSpaceCameraController", function () {
 
   it("camera does not go below the terrain in CV with the transform set", function () {
     setUpCV();
-    scene.globe = new MockGlobe(scene.mapProjection.ellipsoid);
+    scene.globe = new MockGlobe(scene.ellipsoid);
 
     updateController();